# Conflicts:
#	.idea/misc.xml
#	OOP-Arkanoid.iml
#	Project/src/main/java/arkanoid/engine/Ball.java
#	Project/src/main/java/arkanoid/engine/Bricks.java
#	Project/src/main/java/arkanoid/engine/CheckCollision/CheckBallAndBrick.java
#	Project/src/main/java/arkanoid/engine/Define.java
#	Project/src/main/java/arkanoid/engine/GameEngine.java
#	Project/src/main/java/arkanoid/engine/GameObject.java
#	Project/src/main/java/arkanoid/engine/Main.java
#	Project/src/main/java/arkanoid/engine/Map.java
#	Project/src/main/java/arkanoid/engine/PowerUp/Main.java
#	Project/src/main/java/arkanoid/engine/PowerUp/PowerUpType.java
#	Project/src/main/java/com/example/Arkanoid/Data/HighScoreEntry.java
#	Project/src/main/java/com/example/Arkanoid/Data/Lives.java
#	Project/src/main/java/com/example/Arkanoid/Data/Score.java
#	Project/src/main/java/com/example/Arkanoid/Main.java
#	Project/src/main/java/com/example/Arkanoid/UI/EndController.java
#	Project/src/main/java/com/example/Arkanoid/UI/EndScene.java
#	Project/src/main/java/com/example/Arkanoid/UI/GameController.java
#	Project/src/main/java/com/example/Arkanoid/UI/GameScene.java
#	Project/src/main/java/com/example/Arkanoid/UI/GameScene2.java
#	Project/src/main/java/com/example/Arkanoid/UI/HighScoresController.java
#	Project/src/main/java/com/example/Arkanoid/UI/HighScoresScene.java
#	Project/src/main/java/com/example/Arkanoid/UI/LevelController.java
#	Project/src/main/java/com/example/Arkanoid/UI/LevelScene.java
#	Project/src/main/java/com/example/Arkanoid/UI/MenuController.java
#	Project/src/main/java/com/example/Arkanoid/UI/MenuScene.java
#	Project/src/main/java/com/example/Arkanoid/UI/PauseController.java
#	Project/src/main/java/com/example/Arkanoid/UI/PauseScene.java
#	Project/src/main/java/com/example/Arkanoid/UI/SettingController.java
#	Project/src/main/java/com/example/Arkanoid/UI/SettingScene.java
#	Project/src/main/java/com/example/Arkanoid/UI/SpriteLoader.java
#	Project/src/main/java/com/example/Arkanoid/Utlis/Animations/PaddleGlowAnimation.java
#	Project/src/main/java/com/example/Arkanoid/Utlis/SceneCache.java
#	Project/src/main/java/com/example/Arkanoid/Utlis/SceneNavigator.java
#	Project/src/main/java/com/example/Arkanoid/Utlis/SoundManager.java
#	Project/target/classes/com/Arkanoid/EndScreen.fxml
#	Project/target/classes/com/Arkanoid/MenuGame.fxml
#	Project/target/classes/com/Arkanoid/PauseMenu.fxml
#	Project/target/classes/com/Arkanoid/Setting.fxml
#	Project/target/classes/main/java/arkanoid/engine/Bricks.class
#	Project/target/classes/main/java/arkanoid/engine/GameObject.class
#	Project/target/classes/main/java/arkanoid/engine/Map.class
#	Project/target/classes/main/java/arkanoid/engine/Paddle.class
#	Project/target/classes/main/java/com/example/Arkanoid/Data/Score.class
#	Project/target/classes/main/java/com/example/Arkanoid/Main.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/EndController.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/EndScene.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/MenuController.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/MenuScene.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/PauseController.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/PauseScene.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/SettingController.class
#	Project/target/classes/main/java/com/example/Arkanoid/UI/SettingScene.class
#	Project/target/maven-status/maven-compiler-plugin/compile/default-compile/createdFiles.lst
#	Project/target/maven-status/maven-compiler-plugin/compile/default-compile/inputFiles.lst
#	out/production/OOP-Arkanoid/main/java/arkanoid/engine/Ball.class
#	out/production/OOP-Arkanoid/main/java/arkanoid/engine/Bricks.class
#	out/production/OOP-Arkanoid/main/java/arkanoid/engine/GameObject.class
#	out/production/OOP-Arkanoid/main/java/arkanoid/engine/Main.class
#	out/production/OOP-Arkanoid/main/java/arkanoid/engine/Map.class
#	out/production/OOP-Arkanoid/main/java/arkanoid/engine/Paddle.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/Data/Score.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/Main.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/EndController.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/EndScene.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/MenuController.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/MenuScene.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/PauseController.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/PauseScene.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/SettingController.class
#	out/production/OOP-Arkanoid/main/java/com/example/Arkanoid/UI/SettingScene.class
#	out/production/OOP-Arkanoid/main/resources/com/Arkanoid/EndScreen.fxml
#	out/production/OOP-Arkanoid/main/resources/com/Arkanoid/MenuGame.fxml
#	out/production/OOP-Arkanoid/main/resources/com/Arkanoid/PauseMenu.fxml
#	out/production/OOP-Arkanoid/main/resources/com/Arkanoid/Setting.fxml

+++ Project/src/main/java/arkanoid/engine/Bricks.java
@@ -47,7 +47,7 @@ public class Bricks extends GameObject {
         return currrentPoints;
     }
 
-    public void setCurrrentPoints(double minusPoint) {
+    public void setCurrrentPoints(int minusPoint) {
         this.currrentPoints -= minusPoint;
     }
